+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+* Added the `parallel` optional argument to the `Reader` methods
+  `{from_zip, from_dir, from_files, from_strs}`
+  so that parallelization can be turned off if desired.
+
 ### Changed
 ### Deprecated
 ### Removed